requirements.txt
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
django-crispy-forms==1.10.0
django-rest-auth==0.9.5
django-allauth==0.44.0
coreapi==2.3.3
pyyaml==5.3.1
drf-yasg==1.20.0
django-rest-swagger==2.2.0
django-mptt==0.11.0
psycopg2-binary==2.8.6
django-environ==0.4.5
dj-rest-auth==2.1.2






